+++ apps/vscode/vite.config.ts
@@ -4,6 +4,7 @@ import { svelte } from "@sveltejs/vite-plugin-svelte";
 import { defineConfig } from "vite";
 
 export default defineConfig({
+  base: "./",
   plugins: [svelte()],
   resolve: {
     alias: {